Tests/LibWeb/Text/expected/HTML/media-source-detach-on-load.txt

readyState after open: open
sourceBuffers after open: 1
readyState after close: closed
sourceBuffers after close: 0
